incident_management_api/requirements.txt
Django>=4.2,<5
djangorestframework
djangorestframework-simplejwt
drf-yasg
django-filter
mysqlclient
python-dotenv
gunicorn
